refactor: resolve Codex's native title in the CLI hook, not the app
Addresses PR #11171's automated review:

- Move the state_5.sqlite read entirely into a new CodexNativeTitleStore
  in the CMUXAgentLaunch package (already linked by both the CLI and app
  targets, and already home to other Codex session/persistence code like
  CodexSessionResumeVerifier). The detached `cmux hooks codex
  sync-native-title` CLI process now resolves the title itself and sends
  the app the plain string, exactly like workspace.set_auto_title already
  receives a pre-computed title. The app-side handler does zero file/DB
  I/O, so the earlier main-actor-blocking concern is moot by construction
  rather than by threading workaround, and the `#if DEBUG
  db_path_for_testing` socket seam is gone: tests inject dbPath as a
  normal parameter on CodexNativeTitleStore directly.
- Replace spawnDetachedCodexNativeTitleSync's waitForProcessExit(timeout: 1)
  with spawnDetachedCodexSettledStop's existing nohup-and-forget shape —
  no blocking wait on the Stop-hook path at all.
- Fully reverted Sources/SessionIndexStore+CodexSQL.swift and
  cmuxTests/SessionIndexViewTests.swift to their origin/main state; the
  app target now carries none of this persistence logic.
- Rewrote the surface.sync_codex_native_title tests in
  cmuxTests/SetAutoTitleSocketTests.swift to pass a plain title string
  (mirroring workspace.set_auto_title's own test style) and added a test
  proving the sync applies regardless of the opt-in Workspace Auto-Naming
  setting. Added CodexNativeTitleStoreTests.swift in CMUXAgentLaunchTests
  for the reader itself.

Left unchanged, independently verified against the existing codebase
rather than blindly applied: the review's error-string localization
finding. Sources/TerminalController.swift has 303 `.err(code:...)` sites
and zero use String(localized:), including the pre-existing
v2WorkspaceSetAutoTitle this code mirrors — these are internal v2 socket
protocol errors consumed via error.code, not user-facing UI strings.

Fixed after a second independent review pass: a stale comment in
ControlCommandExecutionPolicy.swift/Tests still describing the old
app-side-read architecture, and a test named for an empty-title case it
didn't actually exercise.
