+++ Packages/macOS/CmuxControlSocket/Sources/CmuxControlSocket/Wire/ControlCommandExecutionPolicy.swift
@@ -281,10 +281,10 @@ public enum ControlCommandExecutionPolicy: Sendable, Equatable {
         "notification.create_for_caller",
         "workspace.set_auto_title",
         // Same single-hop worker shape as workspace.set_auto_title just
-        // above, so its SessionIndexStore.codexNativeTitle SQLite read (a
-        // state_5.sqlite snapshot-copy-and-query, modeled on the existing
-        // Vault-search path) runs on the worker thread rather than blocking
-        // the main actor — this fires on every Codex turn end (cmux #11144).
+        // above: the caller (the detached `cmux hooks codex sync-native-title`
+        // process) has already resolved Codex's native title via
+        // CodexNativeTitleStore before sending it, so this handler applies a
+        // plain string with no I/O of its own — cmux #11144.
         "surface.sync_codex_native_title",
         // The v2 resolution reads (tranche D of issue #5757) — the implicit
         // handle-normalization reads nearly every CLI invocation pays 1-3 of.
