+++ docs/workspace-auto-naming.md
@@ -44,7 +44,7 @@ The other hook integrations are intentionally skipped for now:
 
 ## Codex's native tab title (independent of this feature)
 
-Separately from opt-in auto-naming above, cmux mirrors Codex's own natively-generated thread title (`~/.codex/state_5.sqlite`, `threads.title`) onto the panel's raw terminal tab title at every turn end — the same precedence tier OSC terminal-title updates already write to, and any existing custom title (auto or user) is left untouched exactly as it already is for OSC. This runs unconditionally, is not gated by the `automation.workspaceAutoNaming` setting, and never calls a summarizer, because Codex already computed the title itself. It exists because Codex's CLI does not emit OSC title updates tracking its live conversation title the way Claude Code's CLI does, so without this a Codex tab's title would silently lag Claude's (cmux #11144).
+Separately from opt-in auto-naming above, cmux mirrors Codex's own natively-generated thread title (`~/.codex/state_5.sqlite`, `threads.title`, read via the `CodexNativeTitleStore` helper in the `CMUXAgentLaunch` package) onto the panel's raw terminal tab title at every turn end — the same precedence tier OSC terminal-title updates already write to, and any existing custom title (auto or user) is left untouched exactly as it already is for OSC. The detached Codex Stop-hook process resolves the title itself and sends the app the plain string, so the app-side socket handler does no database or file I/O. This runs unconditionally, is not gated by the `automation.workspaceAutoNaming` setting, and never calls a summarizer, because Codex already computed the title itself. It exists because Codex's CLI does not emit OSC title updates tracking its live conversation title the way Claude Code's CLI does, so without this a Codex tab's title would silently lag Claude's (cmux #11144).
 
 ## Mechanics
